requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.34.1
anyio==4.3.0
asgiref==3.8.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
backcall==0.2.0
beautifulsoup4==4.12.3
betterpy==0.5.0
blinker==1.6.2
blis==0.7.11
boto3==1.34.162
botocore==1.34.162
browserbase==0.3.0
bs4==0.0.2
burr==0.28.0
cachetools==5.3.3
camelot-py==0.11.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.6
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.2
confection==0.1.4
cryptography==42.0.5
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.5
dateparser==1.2.0
debugpy==1.6.4
decorator==4.4.2
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dj-database-url==0.5.0
Django==5.1.1
django-cors-headers==4.3.1
django-cron==0.6.0
django-crontab==0.7.1
django-git==0.1
django-heroku==0.3.1
django-mysql==4.8.0
django-paypal==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.6.1
docstring_parser==0.16
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==1.2.0
faiss-cpu==1.8.0.post1
fastapi==0.116.0
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg==1.4
filelock==3.13.4
fireworks-ai==0.15.0
Flask==2.3.2
free_proxy==1.1.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.9
GitPython==3.1.24
google==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.146.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud==0.34.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.2
googleapis-common-protos==1.63.0
GoogleNews==1.6.14
graphviz==0.20.3
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==20.0.4
h11==0.14.0
holidays==0.56
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.4
imageio==2.10.3
imageio-ffmpeg==0.4.5
img2pdf==0.5.1
ipykernel==6.19.3
ipython==8.7.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
jieba3k==0.35.1
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==7.4.8
jupyter_core==5.1.0
kaggle==1.5.12
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
libretranslatepy==2.1.1
load-dotenv==0.1.0
lxml==4.9.2
marisa-trie==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
moviepy==1.0.3
msrest==0.7.1
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nest-asyncio==1.5.6
packaging==25.0
panda==0.3.1
parso==0.8.3
pickleshare==0.7.5
pillow==11.3.0
platformdirs==2.6.0
playwright==1.46.0
pluggy==1.5.0
plum-dispatch==1.7.4
proglog==0.1.9
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.2
pure-eval==0.2.2
pydantic==1.10.22
pydantic-settings==2.5.2
pydantic_core==2.20.1
PyDrive==1.3.1
pyee==11.1.0
Pygments==2.17.2
PyJWT==2.6.0
PyMySQL==1.0.2
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==7.0.0
pytz==2023.3.post1
pywin32==305
PyYAML==6.0.1
pyzmq==24.0.1
q==2.7
razorpay==1.3.0
referencing==0.35.1
regex==2024.4.16
reportlab==4.0.8
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
s3fs==0.4.2
s3transfer==0.10.1
safetensors==0.4.4
schedule==1.1.0
scikit-learn==1.4.2
scipy==1.13.0
scrapegraphai==1.21.0
selenium==4.24.0
semchunk==2.2.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shapely==2.0.6
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Spire.Doc==12.4.0
Spire.Presentation==9.3.0
Spire.Xls==14.4.4
SQLAlchemy==2.0.4
sqlparse==0.4.4
sqlviewer==0.1.1
srsly==2.4.8
stack-data==0.6.2
starlette==0.46.2
sympy==1.13.2
tabulate==0.9.0
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
torch==2.4.0
tornado==6.2
touch==2020.12.3
tqdm==4.66.5
traitlets==5.8.0
transformers==4.44.2
translate==3.6.1
trio==0.25.0
trio-websocket==0.11.1
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
tzlocal==5.2
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.35.0
values==2020.12.3
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.3.4
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.6
whitenoise==5.3.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
yfinance==0.2.43
